# Parallelize duplicate point deletions across segments