--- a/lib/collection/src/collection_manager/holders/segment_holder.rs
+++ b/lib/collection/src/collection_manager/holders/segment_holder.rs
@@ -9,6 +9,7 @@ use std::time::{Duration, Instant};
 
 use common::iterator_ext::IteratorExt;
 use common::tar_ext;
+use futures::future::try_join_all;
 use io::storage_version::StorageVersion;
 use parking_lot::{RwLock, RwLockReadGuard, RwLockUpgradableReadGuard, RwLockWriteGuard};
 use rand::seq::SliceRandom;
@@ -1190,21 +1191,40 @@ impl<'s> SegmentHolder {
     /// If two points have the same id and version, one of them is kept.
     ///
     /// Deduplication works with plain segments only.
-    pub fn deduplicate_points(&self) -> OperationResult<usize> {
+    pub async fn deduplicate_points(&self) -> OperationResult<usize> {
         let points_to_remove = self.find_duplicated_points();
 
-        let mut removed_points = 0;
-        for (segment_id, points) in points_to_remove {
-            let locked_segment = self.get(segment_id).unwrap();
-            let segment_arc = locked_segment.get();
-            let mut write_segment = segment_arc.write();
-            for point_id in points {
-                if let Some(point_version) = write_segment.point_version(point_id) {
-                    removed_points += 1;
-                    write_segment.delete_point(point_version, point_id)?;
-                }
-            }
-        }
+        // Create (blocking) task per segment for points to delete so we can paralellize
+        let tasks = points_to_remove
+            .into_iter()
+            .map(|(segment_id, points)| {
+                let locked_segment = self.get(segment_id).unwrap().clone();
+                tokio::task::spawn_blocking(move || {
+                    let mut removed_points = 0;
+                    let segment_arc = locked_segment.get();
+                    let mut write_segment = segment_arc.write();
+                    for point_id in points {
+                        if let Some(point_version) = write_segment.point_version(point_id) {
+                            removed_points += 1;
+                            write_segment.delete_point(point_version, point_id)?;
+                        }
+                    }
+                    OperationResult::Ok(removed_points)
+                })
+            })
+            .collect::<Vec<_>>();
+
+        // Join and sum results in parallel
+        let removed_points = try_join_all(tasks)
+            .await
+            .map_err(|err| {
+                OperationError::service_error(format!(
+                    "Failed to join task that removes duplicate points from segment: {err}"
+                ))
+            })?
+            .into_iter()
+            .sum::<Result<_, _>>()?;
+
         Ok(removed_points)
     }
 
@@ -1562,8 +1582,8 @@ mod tests {
         );
     }
 
-    #[test]
-    fn test_points_deduplication() {
+    #[tokio::test]
+    async fn test_points_deduplication() {
         let dir = Builder::new().prefix("segment_dir").tempdir().unwrap();
 
         let mut segment1 = build_segment_1(dir.path());
@@ -1588,7 +1608,7 @@ mod tests {
         let sid1 = holder.add_new(segment1);
         let sid2 = holder.add_new(segment2);
 
-        let res = holder.deduplicate_points().unwrap();
+        let res = holder.deduplicate_points().await.unwrap();
 
         assert_eq!(5, res);