Improve performance of dropping duplicate points on start in case there is a lot of them.

This parallelizes the deletions across segments, rather than doing them in sequence.

My test collection has a segment of about 750k points with 5 keyword indices having just 2 distinct values. This segment is duplicated to get 750k duplicates.

This is how long it takes when loading:
- Before PR: 149 seconds
- After PR: 44 seconds

A more extreme case with 5 copies of the above mentioned segment rather than just 2:
- Before PR: 277 seconds
- After PR: 259 seconds
- After PR with [#5072](https://github.com/qdrant/qdrant/pull/5072): 65 seconds

This obviously is an very unlikely scenario, but it's a good way to show the potential benefit.


### All Submissions:

* [x] Contributions should target the `dev` branch. Did you create your branch from `dev`?
* [x] Have you followed the guidelines in our Contributing document?
* [x] Have you checked to ensure there aren't other open [Pull Requests](../../../pulls) for the same update/change?

<!-- You can erase any parts of this template not applicable to your Pull Request. -->

### New Feature Submissions:

1. [x] Does your submission pass tests?
2. [x] Have you formatted your code locally using `cargo +nightly fmt --all` command prior to submission?
3. [x] Have you checked your code using `cargo clippy --all --all-features` command?